Refactor loading functions
